requirements.txt
discord
python-dotenv
requests
yt-dlp
pynacl
beautifulsoup4
sqlite3